Add and rename field in riscli GetRouters() grpc (#260)
In the risserver every bmp server is identified by its address,
specified in the config, but the GetRouters() grpc only presents us with
the name acquired via the BMP session. To clarify the field usage the
router field is renamed to sys_name and an address field is added. The
address field is of type string to match the other grpc methods, which
require a router address to be specified as string.

Co-authored-by: takt <oliver.herms@exaring.de>