Replace paths with same identifier with AddPath
RFC7911 sec 5 par 1 says that the path identifier added by BGP Add Path
uniquely identifies a path for a given prefix. Therefore, when a new
path arrives we need to replace all paths with the same path identifier
for this prefix. Ignoring this leads to routes piling up in the
adjRIBIn.