Abstract logger (#363)
Replace direct to Logrus with a proxy exposing an abstract interface
very close to Logrus.

The changes include:

- add the `util/log` module
- replace the import of logrus with the import of this module
- call `SetLogger` at the main entry point (not strictly necessary)
- use `logrus` explicitely at some point

Also, the BMP server is losing its internal logger. It was used for
tests, but these tests are commented. The test has been updated to use
`SetLogger` instead (but it has not been really tested, ideally, it
should be a text fixture undoing the change once the test has run).

Fix #362.