BMP: separate Listen() and Serve() into two functions
The previous `Listen()` function was more a `ListenAndServe()`
function and it triggers a race condition when calling `Close()`: call
it too early and you get a panic because the listener is not ready.
Just checking for `nil` in `Close()` would trigger another race
condition where the listener would not be stopped despite calling
`Close()` (because it was called too early).